refactor(utils): refactoring tool modules and unifying all import paths
