+++ src/components/config-plugin.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next'
 import { If } from 'react-if-lite'
 import { useStore } from 'valtio-define'
 import { store } from '@/store'
-import { toast } from '@/utils'
+import { toast } from '@/utils/toast'
 import { useDshPlugins } from '../hooks/use-dsh-plugins'
 import { Ellipsis as TextEllipsis } from './ellipsis'
 import { Empty } from './empty'
